# src/config/swagger.yaml
swagger: '2.0'
info:
  description: 'Controle Simples de Postagens com Usuários'
  version: '1.0.0'
  title: 'posts_control'
basePath: '/api/v1'
consumes:
  - 'application/json'
produces:
  - 'application/json'
tags:
  - name: 'Users'
    description: 'User CRUD'
  - name: 'Posts'
    description: 'Post CRUD'
definitions:
  user:
    additionalProperties: false
    type: 'object'
    required:
      - 'full_name'
      - 'user_email'
      - 'user_password'
    properties:
      full_name:
        type: 'string'
      user_email:
        type: 'string'
      user_password:
        type: 'string'
  user_response:
    additionalProperties: false
    type: 'object'
    properties:
      user_id:
        type: 'integer'
      full_name:
        type: 'string'
      user_email:
        type: 'string'
  post:
    additionalProperties: false
    type: 'object'
    required:
      - 'author_id'
      - 'post_text'
    properties:
      post_id:
        type: 'integer'
      author_id:
        type: 'integer'
      post_text:
        type: 'string'
  users_response:
    additionalProperties: false
    type: 'object'
    properties:
      users:
        type: array
        items:
          $ref: '#/definitions/user_response'
  posts_response:
    additionalProperties: false
    type: 'object'
    properties:
      posts:
        type: array
        items:
          $ref: '#/definitions/post'
paths:
  /user:
    get:
      tags:
        - 'Users'
      x-swagger-router-controller: app
      summary: list users
      operationId: listUserHandler
      responses:
        200:
          description: 'users retrieved'
          schema:
            $ref: '#/definitions/users_response'
    post:
      tags:
        - 'Users'
      x-swagger-router-controller: app
      summary: create user
      operationId: createUserHandler
      parameters:
        - in: body
          name: user
          description: create user
          schema:
            type: 'object'
            additionalProperties: false
            required:
              - 'full_name'
              - 'user_email'
              - 'user_password'
            properties:
              full_name:
                type: 'string'
              user_email:
                type: 'string'
              user_password:
                type: 'string'
      responses:
        201:
          description: 'user created'
          schema:
            $ref: '#/definitions/user_response'
  /user/{id}:
    get:
      tags:
        - 'Users'
      x-swagger-router-controller: app
      summary: get user by id
      operationId: listUserByIdHandler
      parameters:
        - name: id
          in: path
          type: integer
          required: true
          description: User ID
      responses:
        200:
          description: 'user retrieved'
          schema:
            $ref: '#/definitions/user_response'
    put:
      tags:
        - 'Users'
      x-swagger-router-controller: app
      summary: update user (full update)
      operationId: updateUserHandler
      parameters:
        - name: id
          in: path
          type: integer
          required: true
          description: User ID
        - in: body
          name: user
          description: update user
          schema:
            additionalProperties: false
            type: 'object'
            required:
              - 'full_name'
              - 'user_email'
              - 'user_password'
            properties:
              full_name:
                type: 'string'
              user_email:
                type: 'string'
              user_password:
                type: 'string'
      responses:
        200:
          description: 'user updated'
          schema:
            $ref: '#/definitions/user_response'
    patch:
      tags:
        - 'Users'
      x-swagger-router-controller: app
      summary: partial update user
      operationId: patchUserHandler
      parameters:
        - name: id
          in: path
          type: integer
          required: true
          description: User ID
        - in: body
          name: user
          description: partial update user
          schema:
            additionalProperties: false
            type: 'object'
            properties:
              full_name:
                type: 'string'
              user_email:
                type: 'string'
              user_password:
                type: 'string'
      responses:
        200:
          description: 'user updated'
          schema:
            $ref: '#/definitions/user_response'
    delete:
      tags:
        - 'Users'
      x-swagger-router-controller: app
      summary: delete user
      operationId: deleteUserHandler
      parameters:
        - name: id
          in: path
          type: integer
          required: true
          description: User ID
      responses:
        204:
          description: 'user deleted'
  /post:
    get:
      tags:
        - 'Posts'
      x-swagger-router-controller: app
      summary: list posts
      operationId: listPostHandler
      parameters:
        - name: user_id
          in: query
          type: integer
          required: false
          description: Filter posts by user ID (optional - if not provided, returns all posts)
      responses:
        200:
          description: 'posts retrieved'
          schema:
            $ref: '#/definitions/posts_response'
    post:
      tags:
        - 'Posts'
      x-swagger-router-controller: app
      summary: create post
      operationId: createPostHandler
      parameters:
        - in: body
          name: post
          description: create post
          schema:
            type: object
            additionalProperties: false
            required:
              - post_text
              - author_id
            properties:
              post_text:
                type: string
              author_id:
                type: integer
      responses:
        201:
          description: 'post created'
          schema:
            $ref: '#/definitions/post'
  /post/{id}:
    put:
      tags:
        - 'Posts'
      x-swagger-router-controller: app
      summary: update post (full update)
      operationId: updatePostHandler
      parameters:
        - name: id
          in: path
          type: integer
          required: true
          description: Post ID
        - in: body
          name: post
          description: update post
          schema:
            additionalProperties: false
            type: 'object'
            required:
              - 'author_id'
              - 'post_text'
            properties:
              author_id:
                type: 'integer'
              post_text:
                type: 'string'
      responses:
        200:
          description: 'post updated'
          schema:
            $ref: '#/definitions/post'
    patch:
      tags:
        - 'Posts'
      x-swagger-router-controller: app
      summary: partial update post
      operationId: patchPostHandler
      parameters:
        - name: id
          in: path
          type: integer
          required: true
          description: Post ID
        - in: body
          name: post
          description: partial update post
          schema:
            additionalProperties: false
            type: 'object'
            properties:
              author_id:
                type: 'integer'
              post_text:
                type: 'string'
      responses:
        200:
          description: 'post updated'
          schema:
            $ref: '#/definitions/post'
    delete:
      tags:
        - 'Posts'
      x-swagger-router-controller: app
      summary: delete post
      operationId: deletePostHandler
      parameters:
        - name: id
          in: path
          type: integer
          required: true
          description: Post ID
      responses:
        204:
          description: 'post deleted'